backend/requirements.txt
flask
torch
torchvision
numpy
scikit-learn
matplotlib
Pillow
